requirements.txt
aiohttp
discord.py
python-dotenv
sqlalchemy
aiosqlite
asyncpg
gspread
google-auth